refresh changes from production at https://wordpress.org/plugins/bestbooks/

+++ class/src/chartofaccounts.php
@@ -5,33 +5,44 @@ class ChartOfAccounts {
    var $count = 0;
 
    public function __construct() {
-   	  global $wpdb;
-   	  
-      if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-        $sql = "SELECT name,type FROM ".$wpdb->base_prefix."bestbooks_accounts";
+      global $wpdb;
+       
+      $sql = '';
+    
+      if (function_exists('is_plugin_active_for_network')) {
+        if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+          $sql = "SELECT name,type FROM ".$wpdb->base_prefix."bestbooks_accounts";
+        } else {
+          $sql = "SELECT name,type FROM ".$wpdb->prefix."bestbooks_accounts";
+        }
+        $result = $wpdb->get_results($sql);
+        foreach($result as $account) {
+           $this->account[$account->name] = $account->type;
+           $this->count++;
+        }  
       } else {
         $sql = "SELECT name,type FROM ".$wpdb->prefix."bestbooks_accounts";
+        $result = $wpdb->get_results($sql);
+        foreach($result as $account) {
+           $this->account[$account->name] = $account->type;
+           $this->count++;
+        }  
       }
-      $result = $wpdb->get_results($sql);
-      foreach($result as $account) {
-         $this->account[$account->name] = $account->type;
-         $this->count++;
-      }
-      //echo '<pre>'; print_r($result); echo '</pre>'; die();
-      //while ($wpdb->get_row($sql,$row,$this->count)) {
-      //   list($name,$type) = $row;
-      //   $this->account[$name] = $type;
-      //   $this->count++;
-      //}
    }
 
    public static function dropTable() {
    	  global $wpdb;
-   	  
-      if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-        $sql = "DROP TABLE ".$wpdb->base_prefix."bestbooks_accounts";
+      
+      $sql = '';
+
+      if (function_exists('is_plugin_active_for_network')) {
+        if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+          $sql = "DROP TABLE ".$wpdb->base_prefix."bestbooks_accounts";
+        } else {
+          $sql = "DROP TABLE ".$wpdb->prefix."bestbooks_accounts";
+        }
       } else {
-        $sql = "DROP TABLE ".$wpdb->prefix."bestbooks_accounts";
+          $sql = "DROP TABLE ".$wpdb->prefix."bestbooks_accounts";
       }
       $result = $wpdb->query($sql);
 
@@ -44,28 +55,41 @@ class ChartOfAccounts {
 
     public static function createTable() {
         global $wpdb;
-   	  
-        //$sql = "CREATE TABLE `".$wpdb->prefix."Accounts` (`name` VARCHAR(50) NOT NULL,`type` VARCHAR(20) NOT NULL,`class` VARCHAR(255) NOT NULL, PRIMARY KEY(`name`))";
-        if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-          $sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->base_prefix."bestbooks_accounts` (
-                          `id` tinyint(4) NOT NULL auto_increment,
-                          `txdate` date NOT NULL default '0000-00-00',
-                          `name` varchar(50) NOT NULL default '',
-                          `type` varchar(20) NOT NULL default '',
-                          `data` varchar(25) NOT NULL default '',
-                          `class` varchar(255) NOT NULL default '',
-                          PRIMARY KEY  (`id`)
-                          ) ENGINE=MyISAM DEFAULT CHARSET=latin1 AUTO_INCREMENT=1";
-      } else {
+       
+        $sql = '';
+
+        if (function_exists('is_plugin_active_for_network')) {
+          if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+            $sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->base_prefix."bestbooks_accounts` (
+                            `id` tinyint(4) NOT NULL auto_increment,
+                            `txdate` date NOT NULL default '0000-00-00',
+                            `name` varchar(50) NOT NULL default '',
+                            `type` varchar(20) NOT NULL default '',
+                            `data` varchar(25) NOT NULL default '',
+                            `class` varchar(255) NOT NULL default '',
+                            PRIMARY KEY  (`id`)
+                            ) ENGINE=MyISAM DEFAULT CHARSET=latin1 AUTO_INCREMENT=1";
+          } else {
+            $sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->prefix."bestbooks_accounts` (
+                            `id` tinyint(4) NOT NULL auto_increment,
+                            `txdate` date NOT NULL default '0000-00-00',
+                            `name` varchar(50) NOT NULL default '',
+                            `type` varchar(20) NOT NULL default '',
+                            `data` varchar(25) NOT NULL default '',
+                            `class` varchar(255) NOT NULL default '',
+                            PRIMARY KEY  (`id`)
+                            ) ENGINE=MyISAM DEFAULT CHARSET=latin1 AUTO_INCREMENT=1";
+          }
+        } else {
           $sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->prefix."bestbooks_accounts` (
-                          `id` tinyint(4) NOT NULL auto_increment,
-                          `txdate` date NOT NULL default '0000-00-00',
-                          `name` varchar(50) NOT NULL default '',
-                          `type` varchar(20) NOT NULL default '',
-                          `data` varchar(25) NOT NULL default '',
-                          `class` varchar(255) NOT NULL default '',
-                          PRIMARY KEY  (`id`)
-                          ) ENGINE=MyISAM DEFAULT CHARSET=latin1 AUTO_INCREMENT=1";
+            `id` tinyint(4) NOT NULL auto_increment,
+            `txdate` date NOT NULL default '0000-00-00',
+            `name` varchar(50) NOT NULL default '',
+            `type` varchar(20) NOT NULL default '',
+            `data` varchar(25) NOT NULL default '',
+            `class` varchar(255) NOT NULL default '',
+            PRIMARY KEY  (`id`)
+            ) ENGINE=MyISAM DEFAULT CHARSET=latin1 AUTO_INCREMENT=1";
         }
         
         $result = $wpdb->query($sql);
@@ -84,9 +108,15 @@ class ChartOfAccounts {
           $this->account[$name] = $type;
           $this->count++;
 
+          $sql = "";
+
           $created = date('Y-m-d');
-          if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-            $sql = "INSERT INTO ".$wpdb->base_prefix."bestbooks_accounts (txdate,name,type) VALUES ('".$created."','".$name."','".$type."')";
+          if (function_exists('is_plugin_active_for_network')) {
+            if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+              $sql = "INSERT INTO ".$wpdb->base_prefix."bestbooks_accounts (txdate,name,type) VALUES ('".$created."','".$name."','".$type."')";
+            } else {
+              $sql = "INSERT INTO ".$wpdb->prefix."bestbooks_accounts (txdate,name,type) VALUES ('".$created."','".$name."','".$type."')";            
+            }
           } else {
             $sql = "INSERT INTO ".$wpdb->prefix."bestbooks_accounts (txdate,name,type) VALUES ('".$created."','".$name."','".$type."')";            
           }
@@ -104,8 +134,13 @@ class ChartOfAccounts {
   function in_use($account) {
     global $wpdb;
 
-    if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-      $sql = "SELECT * FROM ".$wpdb->base_prefix."bestbooks_journal WHERE account='$account'";
+    $sql = '';
+    if (function_exists('is_plugin_active_for_network')) {
+      if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+        $sql = "SELECT * FROM ".$wpdb->base_prefix."bestbooks_journal WHERE account='$account'";
+      } else {
+        $sql = "SELECT * FROM ".$wpdb->prefix."bestbooks_journal WHERE account='$account'";
+      }
     } else {
       $sql = "SELECT * FROM ".$wpdb->prefix."bestbooks_journal WHERE account='$account'";
     }
@@ -120,9 +155,13 @@ class ChartOfAccounts {
 
   function remove($name) {
     global $wpdb;
-
-    if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-      $sql = "DELETE FROM ".$wpdb->base_prefix."bestbooks_accounts WHERE name='$name'";
+    $sql = '';
+    if (function_exists('is_plugin_active_for_network')) {
+      if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+        $sql = "DELETE FROM ".$wpdb->base_prefix."bestbooks_accounts WHERE name='$name'";
+      } else {
+        $sql = "DELETE FROM ".$wpdb->prefix."bestbooks_accounts WHERE name='$name'";
+      }
     } else {
       $sql = "DELETE FROM ".$wpdb->prefix."bestbooks_accounts WHERE name='$name'";
     }
