+++ class/src/ledger.php
@@ -14,11 +14,16 @@ class Ledger extends TAccount {
 			throw new BestBooksException("Null pointer exception");
 		$this->name = $name;
 		$this->type = $type;
-    	if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-			$sql = "SELECT type FROM ".$wpdb->base_prefix."bestbooks_accounts WHERE name='$this->name'";
-    	} else {
+		if (function_exists("is_plugin_active_for_network")) {
+			if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+				$sql = "SELECT type FROM ".$wpdb->base_prefix."bestbooks_accounts WHERE name='$this->name'";
+			} else {
+				$sql = "SELECT type FROM ".$wpdb->prefix."bestbooks_accounts WHERE name='$this->name'";
+			}
+		} else {
 			$sql = "SELECT type FROM ".$wpdb->prefix."bestbooks_accounts WHERE name='$this->name'";
 		}
+	
 		$result = $wpdb->get_results($sql);
 
 		if (!$result) {
@@ -29,14 +34,19 @@ class Ledger extends TAccount {
 		}
 		//$wpdb->get_row($sql,$row,0);
 		$this->type = $result[0]->type;
-
-    	if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-			//$sql = "SELECT Balance FROM ".$wpdb->base_prefix."bestbooks_ledger WHERE name='$name' ORDER BY id DESC";    		
-			$sql = "SELECT SUM(debit)-SUM(credit) AS Balance FROM ".$wpdb->base_prefix."bestbooks_ledger WHERE name='$name'";
-    	} else {
+		if (function_exists("is_plugin_active_for_network")) {
+			if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+				//$sql = "SELECT Balance FROM ".$wpdb->base_prefix."bestbooks_ledger WHERE name='$name' ORDER BY id DESC";    		
+				$sql = "SELECT SUM(debit)-SUM(credit) AS Balance FROM ".$wpdb->base_prefix."bestbooks_ledger WHERE name='$name'";
+			} else {
+				//$sql = "SELECT Balance FROM ".$wpdb->prefix."bestbooks_ledger WHERE name='$name' ORDER BY id DESC";    		
+				$sql = "SELECT SUM(debit)-SUM(credit) AS Balance FROM ".$wpdb->prefix."bestbooks_ledger WHERE name='$name'";
+			}
+		} else {
 			//$sql = "SELECT Balance FROM ".$wpdb->prefix."bestbooks_ledger WHERE name='$name' ORDER BY id DESC";    		
 			$sql = "SELECT SUM(debit)-SUM(credit) AS Balance FROM ".$wpdb->prefix."bestbooks_ledger WHERE name='$name'";
-    	}
+		}
+	
 		$result = $wpdb->get_results($sql);
 		//echo '<pre>'; print_r($result); echo '</pre>';
 		$this->balance = number_format($result[0]->Balance, 2);
@@ -56,14 +66,16 @@ class Ledger extends TAccount {
 
 	public function addDebit($date, $desc, $amount) {
 		global $wpdb;
-        
-        $date = date('Y-m-d', strtotime($date));
-		
-    	if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-			$sql = "INSERT INTO ".$wpdb->base_prefix."bestbooks_ledger (name,txdate,note,debit,balance,type) VALUES ('$this->name','$date','$desc','$amount','$this->balance','$this->type')";    		
-    	} else {
+		if (function_exists("is_plugin_active_for_network")) {
+			if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+				$sql = "INSERT INTO ".$wpdb->base_prefix."bestbooks_ledger (name,txdate,note,debit,balance,type) VALUES ('$this->name','$date','$desc','$amount','$this->balance','$this->type')";    		
+			} else {
+				$sql = "INSERT INTO ".$wpdb->prefix."bestbooks_ledger (name,txdate,note,debit,balance,type) VALUES ('$this->name','$date','$desc','$amount','$this->balance','$this->type')";    		
+			}
+		} else {
 			$sql = "INSERT INTO ".$wpdb->prefix."bestbooks_ledger (name,txdate,note,debit,balance,type) VALUES ('$this->name','$date','$desc','$amount','$this->balance','$this->type')";    		
-    	}
+		}
+			
 		$result = $wpdb->query($sql);
 
 		if ($result===false) {
@@ -80,14 +92,16 @@ class Ledger extends TAccount {
 
 	public function addCredit($date, $desc, $amount) {
 		global $wpdb;
-        
-        $date = date('Y-m-d', strtotime($date));
-		
-    	if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-			$sql = "INSERT INTO ".$wpdb->base_prefix."bestbooks_ledger (name,txdate,note,credit,balance,type) VALUES ('$this->name','$date','$desc','$amount','$this->balance','$this->type')";    		
-    	} else {
+		if (function_exists("is_plugin_active_for_network")) {
+			if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+				$sql = "INSERT INTO ".$wpdb->base_prefix."bestbooks_ledger (name,txdate,note,credit,balance,type) VALUES ('$this->name','$date','$desc','$amount','$this->balance','$this->type')";    		
+			} else {
+				$sql = "INSERT INTO ".$wpdb->prefix."bestbooks_ledger (name,txdate,note,credit,balance,type) VALUES ('$this->name','$date','$desc','$amount','$this->balance','$this->type')";    		
+			}
+		} else {
 			$sql = "INSERT INTO ".$wpdb->prefix."bestbooks_ledger (name,txdate,note,credit,balance,type) VALUES ('$this->name','$date','$desc','$amount','$this->balance','$this->type')";    		
-    	}
+		}
+			
 		$result = $wpdb->query($sql);
 
 		if ($result===false) {
@@ -112,53 +126,60 @@ class Ledger extends TAccount {
 	}
 
 	public function getBalance() {
-        global $wpdb;
-        
-    	if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-			$sql = "SELECT SUM(debit) - SUM(credit) AS Balance FROM ".$wpdb->base_prefix."bestbooks_ledger WHERE name='$this->name'";
-    	} else {
-			$sql = "SELECT SUM(debit) - SUM(credit) AS Balance FROM ".$wpdb->prefix."bestbooks_ledger WHERE name='$this->name'";
-    	}
-		$result = $wpdb->get_results($sql);
-        $this->balance = $result[0]->Balance;
 		return $this->balance;
 	}
 
 	public function setBalance($balance) {
 		$this->balance = $balance;
 	}
-    
+
 	public static function createTable() {
         global $wpdb;
 
-            //$sql = "CREATE TABLE `".$wpdb->prefix."Ledger` (`id` TINYINT AUTO_INCREMENT ,`name` VARCHAR( 255 ) NOT NULL,`txdate` DATE NOT NULL,`desc2` VARCHAR( 255 ) NOT NULL,`ref` DOUBLE NOT NULL,`debit` DECIMAL( 10, 2 ) NOT NULL ,`credit` DECIMAL( 10, 2 ) NOT NULL ,`balance` DECIMAL( 10, 2 ) NOT NULL ,`type` VARCHAR( 10 ) NOT NULL ,PRIMARY KEY ( `id` ) )";
-    	if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-            $sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->base_prefix."bestbooks_ledger` (
-                    `id` tinyint(4) NOT NULL auto_increment,
-                    `name` varchar(255) NOT NULL default '',
-                    `txdate` date NOT NULL default '0000-00-00',
-                    `note` varchar(255) NOT NULL default '',
-                    `ref` double NOT NULL default '0',
-                    `debit` decimal(10,2) NOT NULL default '0.00',
-                    `credit` decimal(10,2) NOT NULL default '0.00',
-                    `balance` decimal(10,2) NOT NULL default '0.00',
-                    `type` varchar(10) NOT NULL default '',
-                    PRIMARY KEY  (`id`)
-                    ) ENGINE=MyISAM  DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;";
-    	} else {
-            $sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->prefix."bestbooks_ledger` (
-                    `id` tinyint(4) NOT NULL auto_increment,
-                    `name` varchar(255) NOT NULL default '',
-                    `txdate` date NOT NULL default '0000-00-00',
-                    `note` varchar(255) NOT NULL default '',
-                    `ref` double NOT NULL default '0',
-                    `debit` decimal(10,2) NOT NULL default '0.00',
-                    `credit` decimal(10,2) NOT NULL default '0.00',
-                    `balance` decimal(10,2) NOT NULL default '0.00',
-                    `type` varchar(10) NOT NULL default '',
-                    PRIMARY KEY  (`id`)
-                    ) ENGINE=MyISAM  DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;";    		
-    	}
+        //$sql = "CREATE TABLE `".$wpdb->prefix."Ledger` (`id` TINYINT AUTO_INCREMENT ,`name` VARCHAR( 255 ) NOT NULL,`txdate` DATE NOT NULL,`desc2` VARCHAR( 255 ) NOT NULL,`ref` DOUBLE NOT NULL,`debit` DECIMAL( 10, 2 ) NOT NULL ,`credit` DECIMAL( 10, 2 ) NOT NULL ,`balance` DECIMAL( 10, 2 ) NOT NULL ,`type` VARCHAR( 10 ) NOT NULL ,PRIMARY KEY ( `id` ) )";
+		if (function_exists("is_plugin_active_for_network")) {
+			if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+				$sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->base_prefix."bestbooks_ledger` (
+						`id` tinyint(4) NOT NULL auto_increment,
+						`name` varchar(255) NOT NULL default '',
+						`txdate` date NOT NULL default '0000-00-00',
+						`note` varchar(255) NOT NULL default '',
+						`ref` double NOT NULL default '0',
+						`debit` decimal(10,2) NOT NULL default '0.00',
+						`credit` decimal(10,2) NOT NULL default '0.00',
+						`balance` decimal(10,2) NOT NULL default '0.00',
+						`type` varchar(10) NOT NULL default '',
+						PRIMARY KEY  (`id`)
+						) ENGINE=MyISAM  DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;";
+			} else {
+				$sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->prefix."bestbooks_ledger` (
+						`id` tinyint(4) NOT NULL auto_increment,
+						`name` varchar(255) NOT NULL default '',
+						`txdate` date NOT NULL default '0000-00-00',
+						`note` varchar(255) NOT NULL default '',
+						`ref` double NOT NULL default '0',
+						`debit` decimal(10,2) NOT NULL default '0.00',
+						`credit` decimal(10,2) NOT NULL default '0.00',
+						`balance` decimal(10,2) NOT NULL default '0.00',
+						`type` varchar(10) NOT NULL default '',
+						PRIMARY KEY  (`id`)
+						) ENGINE=MyISAM  DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;";    		
+			}
+		} else {
+			$sql = "CREATE TABLE IF NOT EXISTS `".$wpdb->prefix."bestbooks_ledger` (
+				`id` tinyint(4) NOT NULL auto_increment,
+				`name` varchar(255) NOT NULL default '',
+				`txdate` date NOT NULL default '0000-00-00',
+				`note` varchar(255) NOT NULL default '',
+				`ref` double NOT NULL default '0',
+				`debit` decimal(10,2) NOT NULL default '0.00',
+				`credit` decimal(10,2) NOT NULL default '0.00',
+				`balance` decimal(10,2) NOT NULL default '0.00',
+				`type` varchar(10) NOT NULL default '',
+				PRIMARY KEY  (`id`)
+				) ENGINE=MyISAM  DEFAULT CHARSET=latin1 AUTO_INCREMENT=1 ;";    		
+}
+	
         $result = $wpdb->query($sql);
 
         if ($result === false) {
@@ -169,13 +190,16 @@ class Ledger extends TAccount {
 	}
 
 	public static function dropTable() {
-		global $wpdb;
-
-		if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-			$sql = "DROP TABLE ".$wpdb->base_prefix."bestbooks_ledger";    		
-    	} else {
+		if (function_exists("is_plugin_active_for_network")) {
+			if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
+				$sql = "DROP TABLE ".$wpdb->base_prefix."bestbooks_ledger";    		
+			} else {
+				$sql = "DROP TABLE ".$wpdb->prefix."bestbooks_ledger";    		
+			}
+		} else {
 			$sql = "DROP TABLE ".$wpdb->prefix."bestbooks_ledger";    		
-    	}
+		}
+	
 		$result = $wpdb->query($sql);
 
 		if ($result===false) {
@@ -183,23 +207,6 @@ class Ledger extends TAccount {
 		}
 
 		return "Ledger table dropped successfully";
-	}
-
-	public function transactions() {
-		global $wpdb;
-
-		if (is_plugin_active_for_network('bestbooks/bestbooks.php')) {
-			$sql = "SELECT * FROM ".$wpdb->base_prefix."bestbooks_ledger WHERE name='$this->name'";    		
-    	} else {
-			$sql = "SELECT * FROM ".$wpdb->prefix."bestbooks_ledger WHERE name='$this->name'";    		
-		}
-		$result = $wpdb->query($sql);
-
-		if ($result===false) {
-			throw new BestBooksException("$this->name transactions retrieval");
-		}
-
-		return $result;
 	}
 }
